refactor: update imports dependencies to news folders
