+++ projects/07-midu-router/src/pages/404.jsx
@@ -1,4 +1,4 @@
-import { Link } from '../Link.jsx'
+import { Link } from '../components/Link'
 
 export default function Page404 () {
   return (
