+++ projects/07-midu-router/src/pages/Home.jsx
@@ -1,4 +1,4 @@
-import { Link } from '../Link.jsx'
+import { Link } from '../components/Link'
 
 export default function HomePage () {
   return (
